V1.3 update. Correct counting down.
Changed the for loop and the way that the LCD information is written to the screen when its counting down. Now it counts down 5 seconds instead of 6 by not displaying the 0th second. The display should now count down properly by resetting the cursor every time it writes the number.